feat(fmt): define fmt disk command for FAT16/32

+++ src/domain/fat.rs
@@ -3,7 +3,7 @@ pub(crate) enum FatValue {
     Free,
     Reserved,
     EndOfChain,
-    Data(u32),
+    Data(u16),
     Bad,
 }
 
@@ -13,7 +13,7 @@ impl From<FatValue> for u16 {
             FatValue::Free => 0x0000,
             FatValue::Reserved => 0x0001,
             FatValue::Bad => 0x0002,
-            FatValue::Data(value) => value as u16,
+            FatValue::Data(value) => value,
             FatValue::EndOfChain => 0xFFFF,
         }
     }
@@ -26,7 +26,7 @@ impl From<u16> for FatValue {
             0x0001 => FatValue::Reserved,
             0x0002 => FatValue::Bad,
             0xFFFF => FatValue::EndOfChain,
-            value => FatValue::Data(value as u32),
+            value => FatValue::Data(value),
         }
     }
 }
